src/main/java/com/example/canierasespringindomain/application/MemberService.java
package com.example.canierasespringindomain.application;

import com.example.canierasespringindomain.domain.model.member.Member;
import com.example.canierasespringindomain.domain.model.member.MemberRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@RequiredArgsConstructor
@Service
public class MemberService {
    private final MemberRepository memberRepository;

    @Transactional
    public Member save(Member member) {
        return memberRepository.save(member);
    }

    @Transactional(readOnly = true)
    public Member findById(Long id) {
        return memberRepository.findById(id);
    }
}
